+++ osu.Game/Screens/Play/StandardHUDOverlay.cs
@@ -72,7 +72,7 @@ namespace osu.Game.Screens.Play
             Margin = new MarginPadding { Top = 20, Right = 10 },
         };
 
-        protected override SettingsDisplay CreateSettingsDisplay() => new SettingsDisplay
+        protected override ReplaySettingsOverlay CreateReplaySettingsOverlay() => new ReplaySettingsOverlay
         {
             Anchor = Anchor.TopRight,
             Origin = Anchor.TopRight,
