+++ osu.Game/osu.Game.csproj
@@ -236,7 +236,7 @@
     <Compile Include="Screens\Play\Settings\CollectionSettings.cs" />
     <Compile Include="Screens\Play\Settings\DiscussionSettings.cs" />
     <Compile Include="Graphics\UserInterface\SettingsContainer.cs" />
-    <Compile Include="Screens\Play\Settings\SettingsDisplay.cs" />
+    <Compile Include="Screens\Play\Settings\ReplaySettingsOverlay.cs" />
     <Compile Include="Screens\Play\Settings\PlaybackSettings.cs" />
     <Compile Include="Screens\Play\PauseContainer.cs" />
     <Compile Include="Screens\Play\SongProgressInfo.cs" />
